+++ Fedora/FedoraGaming.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Get arquitecture
+arq=$(uname -m)
+
 # Ensure the script is run with superuser privileges
 if [[ $EUID -ne 0 ]]; then
    echo "This script must be run as root"
@@ -86,10 +89,14 @@ flatpak install -y flathub net.davidotek.pupgui2 # ProtonUp-Qt
 flatpak install -y flathub one.ablaze.floorp
 
 # Install Heroic Games Launcher
-latest_version=$(curl -sL https://api.github.com/repos/Heroic-Games-Launcher/HeroicGamesLauncher/releases/latest | grep -oP '"tag_name": "\K(.*)(?=")' | sed 's/^v//')
-wget "https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/releases/download/v${latest_version}/heroic-${latest_version}.x86_64.rpm"
-dnf install -y "heroic-${latest_version}.x86_64.rpm"
-rm "heroic-${latest_version}.x86_64.rpm"
+if [ "$arq" = "aarch64" ]; then
+    echo "Your system architecture is not yet compatible with Heroic Games Launcher"
+else 
+    latest_version=$(curl -sL https://api.github.com/repos/Heroic-Games-Launcher/HeroicGamesLauncher/releases/latest | grep -oP '"tag_name": "\K(.*)(?=")' | sed 's/^v//')
+    wget "https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/releases/download/v${latest_version}/heroic-${latest_version}.x86_64.rpm"
+    dnf install -y "heroic-${latest_version}.x86_64.rpm"
+    rm "heroic-${latest_version}.x86_64.rpm"
+fi
 
 # Install prerequisites for EmuDeck
 dnf install -y jq rsync unzip zenity
